// src/app/model/TheoreticalFrameworkByProgramByCompanyType.ts
import { CompanyType } from "./companyType";
import { TheoreticalFramework } from "./theoreticalFramework";
import { Program } from "./program";

export class TheoreticalFrameworkByProgramByCompanyType{
    idTheoreticalFrameworkByProgramByCompanyType: number;
    program: Program;
    companyType: CompanyType;
    theoreticalFramework: TheoreticalFramework;
}